Basic, broken read fn.
We can't take the approach of createing a new UnixStream from the fd
on each iteration (even though it was a terrible idea) - as this will
close the fd after first read, which in turn tears down the tap
device.

Need to deal with the lifetime issues cropping up by making it a
field of the TapDevice.

A problem for another day :)