Rust impl of the TapDevice::alloc function.
This now works as it should. The underlying problem was the the file
descriptor was scoped to the alloc function - boxing this fixed the
issue. It also cleaned up the struct a bit and got rid of
UnixStream.

This also removes all of the C code and build time gcc deps.